requirements.txt
absl-py==0.13.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
idna==3.2
joblib==1.0.1
Markdown==3.3.4
numpy==1.20.2
oauthlib==3.1.1
pandas==1.2.4
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.3
six==1.15.0
TA-Lib==0.4.19
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
tqdm==4.62.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
